+++ Sources/BugsnagPerformance/Private/BugsnagPerformanceSpan+Private.h
@@ -7,13 +7,26 @@
 
 #pragma once
 
-#import <BugsnagPerformance/BugsnagPerformanceSpan.h>
-
-#import "Span.h"
+#import <BugsnagPerformance/BugsnagPerformanceSpanOptions.h>
+#import "BugsnagPerformanceSpanContext+Private.h"
+#import "SpanKind.h"
 
 #import <memory>
 
 using namespace bugsnag;
+
+typedef enum {
+    OnSpanDestroyEnd = 1,
+    OnSpanDestroyAbort = 2,
+} OnSpanDestroyAction;
+
+typedef enum {
+    SpanStateOpen = 1,
+    SpanStateEnded = 2,
+    SpanStateAborted = 3,
+} SpanState;
+
+typedef void (^OnSpanClosed)(BugsnagPerformanceSpan * _Nonnull);
 
 NS_ASSUME_NONNULL_BEGIN
 
@@ -21,15 +34,37 @@ NS_ASSUME_NONNULL_BEGIN
 
 @property(nonatomic, copy) void (^onDumped)(BugsnagPerformanceSpan *);
 
-@property(nonatomic) std::shared_ptr<Span> span;
+@property (nonatomic) CFAbsoluteTime startAbsTime;
+@property (nonatomic) CFAbsoluteTime endAbsTime;
+@property (nonatomic) OnSpanDestroyAction onSpanDestroyAction;
+@property (nonatomic,readwrite) NSString *name;
+@property (nonatomic,readonly) NSMutableDictionary *attributes;
+@property (nonatomic) OnSpanClosed onSpanClosed;
+@property (nonatomic,readwrite) SpanId parentId;
+@property (nonatomic) double samplingProbability;
+@property (nonatomic) BSGFirstClass firstClass;
+@property (nonatomic) SpanKind kind;
 
-- (instancetype)initWithSpan:(std::shared_ptr<bugsnag::Span>)span NS_DESIGNATED_INITIALIZER;
 
-- (void)setAttributes:(NSDictionary *)attributes;
+
+@property(nonatomic) uint64_t startClock;
+
+
+@property(atomic) SpanState state;
+
+- (instancetype)initWithName:(NSString *)name
+                     traceId:(TraceId) traceId
+                      spanId:(SpanId) spanId
+                    parentId:(SpanId) parentId
+                   startTime:(CFAbsoluteTime) startTime
+                  firstClass:(BSGFirstClass) firstClass
+                 onSpanClosed:(OnSpanClosed) onSpanEnded NS_DESIGNATED_INITIALIZER;
+
+- (void)setMultipleAttributes:(NSDictionary *)attributes;
 
 - (BOOL)hasAttribute:(NSString *)attributeName withValue:(id)value;
 
-- (id)getAttribute:(NSString *)attributeName;
+- (_Nullable id)getAttribute:(NSString *)attributeName;
 
 - (void)endWithAbsoluteTime:(CFAbsoluteTime)endTime;
 
@@ -38,6 +73,7 @@ NS_ASSUME_NONNULL_BEGIN
 - (SpanId)parentId;
 - (void)updateName:(NSString *)name;
 - (void)updateStartTime:(NSDate *)startTime;
+- (void)updateSamplingProbability:(double) value;
 
 @end