+++ Sources/BugsnagPerformance/Private/OtlpTraceEncoding.h
@@ -9,8 +9,8 @@
 
 #import <Foundation/Foundation.h>
 
-#import "Span.h"
 #import "OtlpPackage.h"
+#import "BugsnagPerformanceSpan+Private.h"
 
 #import <vector>
 
@@ -21,14 +21,14 @@ public:
     /**
      * Build a package suitable for upload to the backend server.
      */
-    static std::unique_ptr<OtlpPackage> buildUploadPackage(const std::vector<std::shared_ptr<SpanData>> &spans, NSDictionary *resourceAttributes, bool includeSamplingHeader) noexcept;
+    static std::unique_ptr<OtlpPackage> buildUploadPackage(NSArray<BugsnagPerformanceSpan *> *spans, NSDictionary *resourceAttributes, bool includeSamplingHeader) noexcept;
 
     static std::unique_ptr<OtlpPackage> buildPValueRequestPackage() noexcept;
 
 public: // Public for testing only
-    static NSDictionary * encode(const SpanData &span) noexcept;
+    static NSDictionary * encode(BugsnagPerformanceSpan *span) noexcept;
     
-    static NSDictionary * encode(const std::vector<std::shared_ptr<SpanData>> &spans, NSDictionary *resourceAttributes) noexcept;
+    static NSDictionary * encode(NSArray<BugsnagPerformanceSpan *> *spans, NSDictionary *resourceAttributes) noexcept;
     
     static NSArray<NSDictionary *> * encode(NSDictionary *attributes) noexcept;