+++ CHANGELOG.md
@@ -11,6 +11,9 @@ Changelog
 * A fixed `samplingProbability` can now be set in the configuration.
   [300](https://github.com/bugsnag/bugsnag-cocoa-performance/pull/300)
 
+* OnEnd span callbacks are now called on early spans (spans that ended before library start) once the library is started.
+  [298](https://github.com/bugsnag/bugsnag-cocoa-performance/pull/298)
+
 ## 1.7.0 (2024-08-05)
 
 ### Enhancements